test: split phase-era e2e coverage by product area
