+++ tests/e2e/test_schedule_overlaps.py
@@ -177,7 +177,7 @@ def open_unscheduled_action_menu_for_text(page, text):
     return task
 
 
-def test_overlap_warnings_flow():
+def test_schedule_overlaps_flow():
     schedule = build_schedule()
     fixed_now = schedule["FIXED_NOW"]
     T1_TIME = schedule["T1_TIME"]
